+++ docs/source/conf.py
@@ -10,7 +10,7 @@ from datetime import date
 from os.path import abspath, dirname
 from pathlib import Path
 
-from aind_protein_data_transformation import __version__ as package_version
+from aind_hcr_data_transformation import __version__ as package_version
 
 INSTITUTE_NAME = "Allen Institute for Neural Dynamics"
