+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["setuptools", "setuptools-scm"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "aind-protein-data-transformation"
+name = "aind-hcr-data-transformation"
 description = "Generated from aind-library-template"
 license = {text = "MIT"}
 requires-python = ">=3.9.2"
