+++ src/aind_hcr_data_transformation/compress/__init__.py
@@ -10,12 +10,12 @@ __author_emails__ = [
 __license__ = "MIT"
 __description__ = "Repository for transforming CZI data to ome.zarr"
 __url__ = (
-    "https://github.com/AllenNeuralDynamics/aind-protein-data-transformation"
+    "https://github.com/AllenNeuralDynamics/aind-hcr-data-transformation"
 )
 
 __maintainers__ = ["Camilo Laiton"]
 __maintainer_emails__ = [
     "camilo.laiton@alleninstitute.org",
 ]
-__title__ = "aind-protein-data-transformation"
-__status__ = "Development"  # 'Development' 'Production', 'Beta'
+__title__ = "aind-protein-hcr-transformation"
+__status__ = "Production"  # 'Development' 'Production', 'Beta'
