+++ src/aind_hcr_data_transformation/compress/zarr_writer.py
