+++ tests/test_integration.py
@@ -7,10 +7,10 @@ import unittest
 from pathlib import Path
 from unittest.mock import patch
 
-from aind_protein_data_transformation.models import (
+from aind_hcr_data_transformation.models import (
     ZeissJobSettings,
 )
-from aind_protein_data_transformation.zeiss_job import (
+from aind_hcr_data_transformation.zeiss_job import (
     ZeissCompressionJob,
 )
